ACCUMULO-4642 Removed DefaultTBE and made TBE pojo
