+++ server/base/src/test/java/org/apache/accumulo/server/master/balancer/HostRegexTableLoadBalancerTest.java
@@ -51,7 +51,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testInit() {
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     Assert.assertEquals("OOB check interval value is incorrect", 2000, this.getOobCheckMillis());
     Assert.assertFalse(isIpBasedRegex());
     Map<String,Pattern> patterns = this.getPoolNameToRegexPattern();
@@ -74,7 +74,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
   @Test
   public void testBalanceWithMigrations() {
     List<TabletMigration> migrations = new ArrayList<>();
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     long wait = this.balance(Collections.unmodifiableSortedMap(createCurrent(2)), Collections.singleton(new KeyExtent()), migrations);
     Assert.assertEquals(20000, wait);
     Assert.assertEquals(0, migrations.size());
@@ -82,7 +82,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testSplitCurrentByRegexUsingHostname() {
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     Map<String,SortedMap<TServerInstance,TabletServerStatus>> groups = this.splitCurrentByRegex(createCurrent(15));
     Assert.assertEquals(3, groups.size());
     Assert.assertTrue(groups.containsKey(FOO.getTableName()));
@@ -113,7 +113,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testSplitCurrentByRegexUsingOverlappingPools() {
-    init(new DefaultTabletBalancerEnvironment(instance, new TestServerConfigurationFactory(instance) {
+    init(new TabletBalancerEnvironment(instance, new TestServerConfigurationFactory(instance) {
 
       @Override
       public TableConfiguration getTableConfiguration(String tableId) {
@@ -178,7 +178,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testSplitCurrentByRegexUsingIP() {
-    init(new DefaultTabletBalancerEnvironment(instance, new TestServerConfigurationFactory(instance) {
+    init(new TabletBalancerEnvironment(instance, new TestServerConfigurationFactory(instance) {
       @Override
       public synchronized AccumuloConfiguration getSystemConfiguration() {
         HashMap<String,String> props = new HashMap<>();
@@ -244,7 +244,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testAllUnassigned() {
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     Map<KeyExtent,TServerInstance> assignments = new HashMap<>();
     Map<KeyExtent,TServerInstance> unassigned = new HashMap<>();
     for (List<KeyExtent> extents : tableExtents.values()) {
@@ -275,7 +275,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testAllAssigned() {
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     Map<KeyExtent,TServerInstance> assignments = new HashMap<>();
     Map<KeyExtent,TServerInstance> unassigned = new HashMap<>();
     this.getAssignments(Collections.unmodifiableSortedMap(allTabletServers), Collections.unmodifiableMap(unassigned), assignments);
@@ -284,7 +284,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testPartiallyAssigned() {
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     Map<KeyExtent,TServerInstance> assignments = new HashMap<>();
     Map<KeyExtent,TServerInstance> unassigned = new HashMap<>();
     int i = 0;
@@ -319,7 +319,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testUnassignedWithNoTServers() {
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     Map<KeyExtent,TServerInstance> assignments = new HashMap<>();
     Map<KeyExtent,TServerInstance> unassigned = new HashMap<>();
     for (KeyExtent ke : tableExtents.get(BAR.getTableName())) {
@@ -349,7 +349,7 @@ public class HostRegexTableLoadBalancerTest extends BaseHostRegexTableLoadBalanc
 
   @Test
   public void testOutOfBoundsTablets() {
-    init(new DefaultTabletBalancerEnvironment(instance, factory));
+    init(new TabletBalancerEnvironment(instance, factory));
     // Wait to trigger the out of bounds check which will call our version of getOnlineTabletsForTable
     UtilWaitThread.sleep(11000);
     Set<KeyExtent> migrations = new HashSet<>();
