+++ server/master/src/main/java/org/apache/accumulo/master/MasterClientServiceHandler.java
@@ -81,7 +81,7 @@ import org.apache.accumulo.master.tserverOps.ShutdownTServer;
 import org.apache.accumulo.server.client.ClientServiceHandler;
 import org.apache.accumulo.server.master.LiveTServerSet.TServerConnection;
 import org.apache.accumulo.server.master.balancer.DefaultLoadBalancer;
-import org.apache.accumulo.server.master.balancer.DefaultTabletBalancerEnvironment;
+import org.apache.accumulo.server.master.balancer.TabletBalancerEnvironment;
 import org.apache.accumulo.server.master.balancer.TabletBalancer;
 import org.apache.accumulo.server.master.state.TServerInstance;
 import org.apache.accumulo.server.replication.StatusUtil;
@@ -466,7 +466,7 @@ public class MasterClientServiceHandler extends FateServiceHandler implements Ma
     if (property.equals(Property.MASTER_TABLET_BALANCER.getKey())) {
       TabletBalancer balancer = master.getConfiguration().instantiateClassProperty(Property.MASTER_TABLET_BALANCER, TabletBalancer.class,
           new DefaultLoadBalancer());
-      balancer.init(new DefaultTabletBalancerEnvironment(instance, master.getConfigurationFactory()));
+      balancer.init(new TabletBalancerEnvironment(instance, master.getConfigurationFactory()));
       master.tabletBalancer = balancer;
       log.info("tablet balancer changed to " + master.tabletBalancer.getClass().getName());
     }
